pyproject.toml
[project]
name = "vn-news-bot"
version = "0.1.0"
description = "Telegram bot for Vietnam news, weather, and disaster alerts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7.3",
]

[project.scripts]
vn-news-bot = "vn_news_bot.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.13",
    "ruff>=0.8",
    "types-pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["feedparser", "feedparser.*"]
ignore_missing_imports = true
follow_untyped_imports = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
